+++ packages/esm-patient-hiv-art-app/src/home-dashboard/charts/DueForViralLoad.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from "react";
 import { LineChart } from "@carbon/charts-react";
 import "@carbon/charts-react/styles.css";
-import styles from "./index.scss";
+import styles from "./styles/index.scss";
 import { DashboardContext } from "../context/DashboardContext";
 import { Loading } from "@carbon/react";
 import { ScaleTypes } from "../types";
