+++ packages/esm-patient-hiv-art-app/src/home-dashboard/charts/ViralLoadSamples.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from "react";
 import { LineChart } from "@carbon/charts-react";
 import "@carbon/charts-react/styles.css";
-import styles from "./index.scss";
+import styles from "./styles/index.scss";
 import { Loading } from "@carbon/react";
 import { DashboardContext } from "../context/DashboardContext";
 import { ScaleTypes } from "../types";
